+++ packages/client/src/devices/InputMediaDeviceManagerState.ts
@@ -7,7 +7,6 @@ import {
 import { isReactNative } from '../helpers/platforms';
 import { RxUtils } from '../store';
 import { getLogger } from '../logger';
-import { isSafari } from '../helpers/browsers';
 
 export type InputDeviceStatus = 'enabled' | 'disabled' | undefined;
 
@@ -67,15 +66,15 @@ export abstract class InputMediaDeviceManagerState<C = MediaTrackConstraints> {
     }
 
     let permissionState: PermissionStatus;
-    const notify = () =>
+    const notify = () => {
       subscriber.next(
-        // In Safari, the `change` event doesn't reliably emit and hence,
+        // In some browsers, the 'change' event doesn't reliably emit and hence,
         // permissionState stays in 'prompt' state forever.
+        // Typically, this happens when a user grants one-time permission.
         // Instead of checking if a permission is granted, we check if it isn't denied
-        isSafari()
-          ? permissionState.state !== 'denied'
-          : permissionState.state === 'granted',
+        permissionState.state !== 'denied',
       );
+    };
     navigator.permissions
       .query({ name: this.permissionName })
       .then((permissionStatus) => {