+++ sample-apps/react/react-dogfood/components/ToggleMicButton.tsx
@@ -10,29 +10,28 @@ import {
   useI18n,
 } from '@stream-io/video-react-sdk';
 
-export const ToggleMenuButton = forwardRef<
-  HTMLButtonElement,
-  ToggleMenuButtonProps
->(function ToggleMenuButton(props, ref) {
-  const { t } = useI18n();
-  const { useMicrophoneState } = useCallStateHooks();
-  const { selectedDevice: selectedMic, devices: microphones } =
-    useMicrophoneState();
+const ToggleMenuButton = forwardRef<HTMLButtonElement, ToggleMenuButtonProps>(
+  function ToggleMenuButton(props, ref) {
+    const { t } = useI18n();
+    const { useMicrophoneState } = useCallStateHooks();
+    const { selectedDevice: selectedMic, devices: microphones } =
+      useMicrophoneState();
 
-  return (
-    <button
-      ref={ref}
-      className="rd__button rd__button--align-left rd__lobby__mic-button"
-    >
-      <Icon className="rd__button__icon" icon="mic" />
-      <p className="rd__lobby__mic-button__device">
-        {microphones?.find((mic) => mic.deviceId === selectedMic)?.label ||
-          t('Default')}
-      </p>
-      <Icon icon={props.menuShown ? 'chevron-down' : 'chevron-up'} />
-    </button>
-  );
-});
+    return (
+      <button
+        ref={ref}
+        className="rd__button rd__button--align-left rd__lobby__mic-button"
+      >
+        <Icon className="rd__button__icon" icon="mic" />
+        <p className="rd__lobby__mic-button__device">
+          {microphones?.find((mic) => mic.deviceId === selectedMic)?.label ||
+            t('Default')}
+        </p>
+        <Icon icon={props.menuShown ? 'chevron-down' : 'chevron-up'} />
+      </button>
+    );
+  },
+);
 
 export const ToggleMicButton = () => {
   return (