+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,47 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.5.17](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.16...@stream-io/video-react-native-sdk-0.5.17) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.10`
+* `@stream-io/video-react-bindings` updated to version `0.4.10`
+### [0.5.16](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.15...@stream-io/video-react-native-sdk-0.5.16) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.9`
+* `@stream-io/video-react-bindings` updated to version `0.4.9`
+### [0.5.15](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.14...@stream-io/video-react-native-sdk-0.5.15) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.8`
+* `@stream-io/video-react-bindings` updated to version `0.4.8`
+
+### Features
+
+* **react:** Support for Background Filters and Background Blurring ([#1283](https://github.com/GetStream/stream-video-js/issues/1283)) ([f790ee7](https://github.com/GetStream/stream-video-js/commit/f790ee78c20fb0f5266e429a777d8bb7ef158c83)), closes [#1271](https://github.com/GetStream/stream-video-js/issues/1271) [#1276](https://github.com/GetStream/stream-video-js/issues/1276)
+
+### [0.5.14](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.13...@stream-io/video-react-native-sdk-0.5.14) (2024-04-02)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.7`
+* `@stream-io/video-react-bindings` updated to version `0.4.7`
+### [0.5.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.12...@stream-io/video-react-native-sdk-0.5.13) (2024-03-29)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.6`
+* `@stream-io/video-react-bindings` updated to version `0.4.6`
+### [0.5.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.11...@stream-io/video-react-native-sdk-0.5.12) (2024-03-29)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.5`
+* `@stream-io/video-react-bindings` updated to version `0.4.5`
 ### [0.5.11](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.5.10...@stream-io/video-react-native-sdk-0.5.11) (2024-03-28)
 
 ### Dependency Updates